search_algos v0.08

+++ 0x1E-search_algorithms/0-linear.c
@@ -21,12 +21,11 @@ int linear_search(int *array, size_t size, int value)
 	}
 	for (; i < size; i++)
 	{
-		printf("Value checked array[%llu] = [%d]\n", i, array[i]);
+		printf("Value checked array[%lu] = [%d]\n", (unsigned long) i, array[i]);
 		if (array[i] == value)
 		{
 			return (i);
 		}
 	}
-
 	return (-1);
 }
